[backend] move make bucket back to docker-compose
